Remove heavy includes from Fiber.hh
